website_backend/requirements.txt
Flask>=2.2.0
Flask-SocketIO>=5.3.0
python-socketio>=5.7.0
python-engineio>=4.3.0
eventlet>=0.33.0
kafka-python>=2.0.0